[example/https] Enable certificate verification
Previously, the self-signed certificate is generated for the server on
the fly, the client has to disable server SSL certificate verification
in order to work with that. This commit adds a self-signed certificate
that the server can load and use it while the client can load it and
trust it.

In addition, the client/server host:port defaults to localhost:8443, to
minimize work with command line while running/debugging the code. Also,
some code cleanups are made.