Added SNI support for the https example server
This is essentially a demo usage of the SniHandler, which allows the
same host+port to use different certificates for different virtual
servers.

Changes to the https server part:
0) A predefined certificate is used by default for all requests, while
another self-assigned certificate is created in runtime for host strings
of format '*.self-assigned.cn'.
1) A SniHandler is created from the <host-string, ssl-context> mapping,
added to the pipeline in HttpsServerInitializer#initChannel(), replacing
the old ssl handler.
2) The HttpsServerHandler#channelRead() use
HttpRequest.headers().get("Host") to fetch the host string used in
request.

Client-side support is not yet available for Netty, but any decent HTTPS
client (browsers, httpie, etc.) shall work.